// apps/app/src/features/questionnaire/interfaces/user-info.ts
export const UserType = { admin: 'admin', general: 'general', guest: 'guest' } as const;

type guestType = typeof UserType.guest;
export type UserType = (typeof UserType)[keyof typeof UserType];

export type IUserInfo =
  | {
      userIdHash: string; // userId hash generated by using appSiteUrl as salt
      type: Exclude<UserType, guestType>;
      userCreatedAt: Date; // createdAt of user that answered the questionnaire
    }
  | {
      type: guestType;
    };